Ghost guest creator

+++ server/controllers/users.js
@@ -9,17 +9,23 @@ userRouter.get('/', async (request, response) => {
         .populate('myEvents', { _id: 1, label: 1, background: 1 })
         .populate('myInvites', { _id: 1, label: 1, background: 1 })
 
-    response.json(users)
+    response.json(users.map(User.format))
 })
 
-userRouter.get('/:id', async (request, response) => {
+userRouter.get('/:id', middleware.requireAuthentication, async (request, response) => {
     try {
         const user = await User
             .findById(request.params.id)
             .populate('myEvents', { _id: 1, label: 1, background: 1 })
             .populate('myInvites', { _id: 1, label: 1, background: 1 })
 
-        response.json(User.format(user))
+        const userId = request.senderId
+
+        if (userId === request.params.id) {
+            response.json(User.format(user))
+        } else {
+            response.json(User.formatForGuest(user))
+        }
     } catch (exception) {
         response.status(400).send({ error: 'Malformatted id' })
     }
@@ -71,6 +77,12 @@ userRouter.put('/:id', middleware.requireAuthentication, async (request, respons
     try {
         const body = request.body
 
+        const userId = request.senderId
+
+        if (userId !== request.params.id) {
+            return response.status(403).send({ error: 'only user itself can update' })
+        }
+
         const existingUser = await User.find({ username: body.username })
         if (existingUser.length > 0) {
             return response.status(400).json({ error: 'username must be unique' })
@@ -109,6 +121,12 @@ userRouter.put('/:id', middleware.requireAuthentication, async (request, respons
 
 userRouter.delete('/:id', middleware.requireAuthentication, async (request, response) => {
     try {
+        const userId = request.senderId
+
+        if (userId !== request.params.id) {
+            return response.status(403).send({ error: 'only user itself can delete' })
+        }
+
         await User.findByIdAndDelete(request.params.id)
         
         response.status(204).end()
